CRUD voltooid

+++ personaliagegevens.php
@@ -3,7 +3,8 @@
 	
 	$query = "SELECT * FROM `personalia`";
 	
-	$result = mysqli_query($conn, $query) or die("query-fout: ".$query." De MySQl-server meldt het volgende: ".mysqli_error($conn)); 	 
+	$result = mysqli_query($conn, $query) or die("query-fout: ".$query." De MySQl-server meldt het volgende: ".mysqli_error($conn)); 
+	
 ?>
 
 <!DOCTYPE html>
@@ -31,45 +32,53 @@ html, table
 		
 <?php 
 		$result_array = mysqli_fetch_assoc($result);
-		echo "<tr>
-					<th>".array_keys($result_array)[0]."</th>
-					<th>".array_keys($result_array)[1]."</th>
-					<th>".array_keys($result_array)[2]."</th>
-					<th>".array_keys($result_array)[3]."</th>
-					<th>".array_keys($result_array)[4]."</th>
-					<th>".array_keys($result_array)[5]."</th>
-					<th>".array_keys($result_array)[6]."</th>
-					<th>".array_keys($result_array)[7]."</th>
-					<th>".array_keys($result_array)[8]."</th>
-					<th></th>
-					<th></th>
-		  	 </tr>";
-			 
-		mysqli_data_seek($result, 0);
 		
-		while ( $result_array = mysqli_fetch_assoc($result) )
+		if ($result_array == null )
+		{
+			echo "<p>Er zijn geen records gevonden in de tabel personalia</p>";
+		}
+		else
 		{
 			echo "<tr>
-					<td>".$result_array["id"]."</td>
-					<td>".$result_array["voornaam"]."</td>
-					<td>".$result_array["tussenvoegsel"]."</td>
-					<td>".$result_array["achternaam"]."</td>
-					<td>".$result_array["email"]."</td>
-					<td>".$result_array["wachtwoord"]."</td>
-					<td>".$result_array["leeftijd"]."</td>
-					<td>".$result_array["haarkleur"]."</td>
-					<td>".$result_array["geboortedatum"]."</td>
-					<td>
-						<a href='update_personalia.php?id=".$result_array["id"]."'>
-							<img src='icons/edit.png' alt='potlood'>
-						</a>
-					</td>
-					<td>
-						<a href='delete_personalia.php?id=".$result_array["id"]."'>
-							<img src='icons/drop.png' alt='delete'>
-						</a>
-					</td>
-				</tr>";
+						<th>".array_keys($result_array)[0]."</th>
+						<th>".array_keys($result_array)[1]."</th>
+						<th>".array_keys($result_array)[2]."</th>
+						<th>".array_keys($result_array)[3]."</th>
+						<th>".array_keys($result_array)[4]."</th>
+						<th>".array_keys($result_array)[5]."</th>
+						<th>".array_keys($result_array)[6]."</th>
+						<th>".array_keys($result_array)[7]."</th>
+						<th>".array_keys($result_array)[8]."</th>
+						<th></th>
+						<th></th>
+				 </tr>";
+				 
+			mysqli_data_seek($result, 0);
+			
+			while ( $result_array = mysqli_fetch_assoc($result) )
+			{
+				echo "<tr>
+						<td>".$result_array["id"]."</td>
+						<td>".$result_array["voornaam"]."</td>
+						<td>".$result_array["tussenvoegsel"]."</td>
+						<td>".$result_array["achternaam"]."</td>
+						<td>".$result_array["email"]."</td>
+						<td>".$result_array["wachtwoord"]."</td>
+						<td>".$result_array["leeftijd"]."</td>
+						<td>".$result_array["haarkleur"]."</td>
+						<td>".$result_array["geboortedatum"]."</td>
+						<td>
+							<a href='update_personalia.php?id=".$result_array["id"]."'>
+								<img src='icons/edit.png' alt='potlood'>
+							</a>
+						</td>
+						<td>
+							<a href='delete_personalia.php?id=".$result_array["id"]."'>
+								<img src='icons/drop.png' alt='delete'>
+							</a>
+						</td>
+					</tr>";
+			}
 		}
 ?>
 	</table>
